+++ packages/netlify-cms-backend-github/src/AuthenticationPage.js
@@ -9,6 +9,18 @@ const LoginButtonIcon = styled(Icon)`
   margin-right: 18px;
 `;
 
+const ForkApprovalContainer = styled.div`
+  display: flex;
+  flex-flow: column nowrap;
+  justify-content: space-around;
+  flex-grow: 0.2;
+`;
+const ForkButtonsContainer = styled.div`
+  display: flex;
+  flex-flow: row nowrap;
+  justify-content: space-around;
+`;
+
 export default class GitHubAuthenticationPage extends React.Component {
   static propTypes = {
     onLogin: PropTypes.func.isRequired,
@@ -21,6 +33,35 @@ export default class GitHubAuthenticationPage extends React.Component {
   };
 
   state = {};
+
+  getPermissionToFork = () => {
+    return new Promise((resolve, reject) => {
+      this.setState({
+        requestingFork: true,
+        approveFork: () => {
+          this.setState({ requestingFork: false });
+          resolve();
+        },
+        refuseFork: () => {
+          this.setState({ requestingFork: false });
+          reject();
+        },
+      });
+    });
+  };
+
+  loginWithOpenAuthoring(data) {
+    const { backend } = this.props;
+
+    this.setState({ findingFork: true });
+    return backend
+      .authenticateWithFork({ userData: data, getPermissionToFork: this.getPermissionToFork })
+      .catch(err => {
+        this.setState({ findingFork: false });
+        console.error(err);
+        throw err;
+      });
+  }
 
   handleLogin = e => {
     e.preventDefault();
@@ -39,23 +80,60 @@ export default class GitHubAuthenticationPage extends React.Component {
         this.setState({ loginError: err.toString() });
         return;
       }
+      if (this.props.config.getIn(['backend', 'open_authoring'])) {
+        return this.loginWithOpenAuthoring(data).then(() => this.props.onLogin(data));
+      }
       this.props.onLogin(data);
     });
   };
 
+  renderLoginButton = () =>
+    this.props.inProgress || this.state.findingFork ? (
+      'Logging in...'
+    ) : (
+      <React.Fragment>
+        <LoginButtonIcon type="github" />
+        {' Login with GitHub'}
+      </React.Fragment>
+    );
+
+  getAuthenticationPageRenderArgs() {
+    const { requestingFork } = this.state;
+
+    if (requestingFork) {
+      const { approveFork, refuseFork } = this.state;
+      return {
+        renderPageContent: ({ LoginButton }) => (
+          <ForkApprovalContainer>
+            <p>
+              Open Authoring is enabled: we need to use a fork on your github account. (If a fork
+              already exists, we&#39;ll use that.)
+            </p>
+            <ForkButtonsContainer>
+              <LoginButton onClick={approveFork}>Fork the repo</LoginButton>
+              <LoginButton onClick={refuseFork}>Don&#39;t fork the repo</LoginButton>
+            </ForkButtonsContainer>
+          </ForkApprovalContainer>
+        ),
+      };
+    }
+
+    return {
+      renderButtonContent: this.renderLoginButton,
+    };
+  }
+
   render() {
     const { inProgress, config } = this.props;
+    const { loginError, requestingFork, findingFork } = this.state;
+
     return (
       <AuthenticationPage
         onLogin={this.handleLogin}
-        loginDisabled={inProgress}
-        loginErrorMessage={this.state.loginError}
+        loginDisabled={inProgress || findingFork || requestingFork}
+        loginErrorMessage={loginError}
         logoUrl={config.get('logo_url')}
-        renderButtonContent={() => (
-          <React.Fragment>
-            <LoginButtonIcon type="github" /> {inProgress ? 'Logging in...' : 'Login with GitHub'}
-          </React.Fragment>
-        )}
+        {...this.getAuthenticationPageRenderArgs()}
       />
     );
   }