+++ packages/netlify-cms-lib-auth/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.4](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-lib-auth/compare/netlify-cms-lib-auth@2.2.3...netlify-cms-lib-auth@2.2.4) (2019-07-24)
+
+**Note:** Version bump only for package netlify-cms-lib-auth
+
+
+
+
+
+## [2.2.3](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-lib-auth/compare/netlify-cms-lib-auth@2.2.2...netlify-cms-lib-auth@2.2.3) (2019-06-26)
+
+**Note:** Version bump only for package netlify-cms-lib-auth
+
+
+
+
+
 ## [2.2.2](https://github.com/netlify/netlify-cms/tree/master/packages/netlify-cms-lib-auth/compare/netlify-cms-lib-auth@2.2.2-beta.0...netlify-cms-lib-auth@2.2.2) (2019-04-10)
 
 **Note:** Version bump only for package netlify-cms-lib-auth