+++ packages/netlify-cms/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.9.8-beta.0](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.9.7...netlify-cms@2.9.8-beta.0) (2019-07-24)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.9.7](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.9.6...netlify-cms@2.9.7) (2019-07-24)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.9.6](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.9.5...netlify-cms@2.9.6) (2019-07-11)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.9.5](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.9.4...netlify-cms@2.9.5) (2019-06-28)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.9.4](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.9.3...netlify-cms@2.9.4) (2019-06-26)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
 ## [2.9.3](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.9.2...netlify-cms@2.9.3) (2019-06-18)
 
 **Note:** Version bump only for package netlify-cms